+++ lib/MonkeyMan/Parameters.pm
@@ -9,6 +9,8 @@ MonkeyMan::Parameters - options passed to a MonkeyMan-driven application
 use strict;
 use warnings;
 
+use MonkeyMan::Exception qw(SuperflousParametersGiven);
+
 # Use Moose and be happy :)
 use Moose;
 use namespace::autoclean;
@@ -45,7 +47,7 @@ method _build_parameters {
 
 method BUILD(...) {
 
-    my $mm              = $self->get_monkeyman;
+    my $monkeyman       = $self->get_monkeyman;
     my $parameters_got  = $self->_get_parameters;
 
     # Parsing parameters
@@ -54,7 +56,7 @@ method BUILD(...) {
         my(
             $parameter_definition,
             $parameter_name
-        ) = each(%{$mm->_get_parameters_to_get})
+        ) = each(%{$monkeyman->_get_parameters_to_get})
     ) {
         $parameters{$parameter_definition} = \($parameters_got->{$parameter_name});
     }
@@ -64,23 +66,73 @@ method BUILD(...) {
     # Adding methods
     my $meta = $self->meta;
     foreach my $parameter_name (keys(%{$parameters_got})) {
+
+        my $predicate =  "has_$parameter_name";
         my $reader    =  "get_$parameter_name";
         my $writer    = "_set_$parameter_name";
-        my $predicate =  "has_$parameter_name";
+
         $meta->add_attribute(
             Class::MOP::Attribute->new(
                 $parameter_name => (
+                    predicate   => $predicate,
                     reader      => $reader,
                     writer      => $writer,
-                    predicate   => $predicate,
                     is          => 'ro'
                 )
             )
         );
-        $meta->add_method(
-            $parameter_name => sub { shift->$reader(@_); }
-        );
+
         $self->$writer($parameters_got->{$parameter_name});
+
+        # Actually we don't need it:
+        #
+        # $meta->add_method(
+        #     $parameter_name => sub { shift->$reader(@_); }
+        # );
+    }
+
+}
+
+
+
+method only_one(Bool :$fatal, ArrayRef[Str] :$lone_attributes) {
+
+    my $monkeyman                   = $self->get_monkeyman;
+    my $logger                      = $monkeyman->get_logger;
+    my %parameters_by_definitions   = %{ $self->get_monkeyman->_get_parameters_to_get };
+    my %parameters_by_attributes    = reverse(%parameters_by_definitions);
+    my @result;
+
+    foreach my $lone_attribute (@{ $lone_attributes }) {
+        my $reader  = 'get_' . $lone_attribute;
+        my $value   = $self->$reader;
+        if(defined($value)) {
+            push(@result, {
+                attribute   => $lone_attribute,
+                definition  => $parameters_by_attributes{$lone_attribute},
+                value       => $value
+            });
+        }
+    }
+
+    if(@result > 1) {
+        my @superflous_parameters;
+        for my $i (2..@result) {
+            push(@superflous_parameters, $result[$i-1]->{'attribute'});
+            $logger->warnf(
+                "The %s parameter (%s) is already given, " .
+                "the %s parameter (%s) is superflous",
+                $result[0]   ->{'attribute'}, $result[0]   ->{'definition'},
+                $result[$i-1]->{'attribute'}, $result[$i-1]->{'definition'}
+            );
+        }
+        if($fatal) {
+            (__PACKAGE__ . '::Exception::SuperflousParametersGiven')->throwf(
+                "Superflous parameter(s) found (%s), " .
+                "which is considered as fatal for this application",
+                join(', ', @superflous_parameters)
+            );
+        }
     }
 
 }